SensorReader/requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-bme680==3.7.12
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-sgp30==3.0.11
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
board==1.0
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
smbus2==0.5.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
pytest-mock
websockets==15.0.1
asyncio==3.4.3